Moved TTT information from MainActivity.java to TTTFragment.java. Only delegates for the two clickhandlers exist in MainActivity.
