+++ app/src/main/java/com/pajato/android/gamechat/fragment/GameFragment.java
@@ -24,7 +24,7 @@ public class GameFragment extends Fragment{
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_game, container, false);
+        return inflater.inflate(R.layout.fragment_ttt, container, false);
     }
 
 }
