+++ app/src/main/java/com/pajato/android/gamechat/fragment/TTTFragment.java
@@ -11,6 +11,7 @@ import android.widget.TextView;
 
 import com.pajato.android.gamechat.R;
 
+import java.util.ArrayList;
 import java.util.Scanner;
 
 public class TTTFragment extends Fragment {
@@ -25,6 +26,18 @@ public class TTTFragment extends Fragment {
     // The mBoard view
     private View mBoard;
 
+    // The Match History
+    private ArrayList<String> mInstructions;
+    private static final String mKEY_INSTRUCTIONS = "match_history";
+
+    // Keeps track of the mTurn user. True = Player 1, False = Player 2.
+    private boolean mTurn;
+    private static final String mKEY_TURN = "current_turn";
+
+    // Player Turn Strings
+    private String mPLAYER1;
+    private String mPLAYER2;
+
     public TTTFragment() {
 
     }
@@ -34,14 +47,85 @@ public class TTTFragment extends Fragment {
         mXValue = getString(R.string.xValue);
         mOValue = getString(R.string.oValue);
 
+        // Initialize the member variables.
+        mPLAYER1 = getString(R.string.player_1);
+        mPLAYER2 = getString(R.string.player_2);
+        mInstructions = new ArrayList<>();
+        mTurn = true;
+
         // Initialize the turn counter, and mBoard variables.
         mTurnCount = 9;
-        mBoard = inflater.inflate(R.layout.fragment_game, container, false);
+        mBoard = inflater.inflate(R.layout.fragment_ttt, container, false);
 
         TextView turnDisplay = (TextView) mBoard.findViewById(R.id.turnDisplay);
         turnDisplay.setText(getTurn(true));
 
         return mBoard;
+    }
+
+    /**
+     * Sends a message alerting the event handling system that the new game button was clicked.
+     *
+     * @param view the new game button.
+     */
+    public void onNewGame(final View view) {
+        // Create the message.
+        String msg = getTurn() + "\n";
+        msg += view.getTag().toString();
+
+        // Empty the instructions list, as a new game has begun.
+        mInstructions.clear();
+        mInstructions.add(msg);
+
+        // Output New Game Message
+        String newTurn = (getTurn().equals(getString(R.string.player_1)) ?
+                "Player 1 (" + getString(R.string.xValue) + ")" :
+                "Player 2 (" + getString(R.string.oValue) + ")") + "'s Turn";
+        Snackbar start = Snackbar.make(getActivity().findViewById(R.id.activity_main), "New Game! " + newTurn, Snackbar.LENGTH_SHORT);
+        start.show();
+
+        //TODO: replace this with an implemented event handling system.
+        sendMessage(msg);
+    }
+
+    /**
+     * Sends a message alerting the event handling system that there was a tile clicked, and
+     * swaps the mTurn to the opposite player.
+     *
+     * @param view the tile clicked
+     */
+    public void tileOnClick(final View view) {
+        String msg = getTurn() + "\n";
+        msg = msg + view.getTag().toString();
+
+        // Keep track of mInstructions for recreating the board.
+        mInstructions.add(msg);
+
+        //TODO: replace this with an implemented event handling system.
+        sendMessage(msg);
+
+        mTurn = !mTurn;
+    }
+
+    /**
+     * A placeholder method for a message handler / event coordinator to be implemented at a later time.
+     *
+     * @param msg the message to transmit to the message handler.
+     */
+    private void sendMessage(final String msg) {
+        //TODO: replace this with an implemented event handling system.
+        // This will be a switch for each of the individual game fragment handlers.
+        messageHandler(msg);
+    }
+
+    /**
+     * Gets the current mTurn and returns a string reflecting the player's
+     * name who is currently playing.
+     *
+     * @return player 1 or player 2, depending on the mTurn.
+     */
+    private String getTurn() {
+        return mTurn ? mPLAYER1 : mPLAYER2;
     }
 
     /**
