+++ app/src/main/java/com/pajato/android/gamechat/main/MainActivity.java
@@ -62,21 +62,6 @@ public class MainActivity extends AppCompatActivity
 
     // Private instance variables
 
-    // The Match History
-    private ArrayList<String> mInstructions;
-    private static final String mKEY_INSTRUCTIONS = "match_history";
-
-    // Keeps track of the mTurn user. True = Player 1, False = Player 2.
-    private boolean mTurn;
-    private static final String mKEY_TURN = "current_turn";
-
-    // Player Turn Strings
-    private String mPLAYER1;
-    private String mPLAYER2;
-
-    // Individual Game Fragments
-    private TTTFragment ticTacToe;
-
     // Public instance methods
 
     // Protected instance methods
@@ -121,37 +106,6 @@ public class MainActivity extends AppCompatActivity
             //startActivity(introIntent);
             //finish();
         }
-
-        // Initialize the member variables.
-        mPLAYER1 = getString(R.string.player_1);
-        mPLAYER2 = getString(R.string.player_2);
-        mInstructions = new ArrayList<>();
-        mTurn = true;
-
-        // Setup the Tic-Tac-Toe fragment and inflate its layout.
-        ticTacToe = (TTTFragment) Panel.game.getFragment(getApplicationContext());
-        ticTacToe.setArguments(getIntent().getExtras());
-        /*
-        getSupportFragmentManager()
-                .beginTransaction()
-                .add(R.id.fragment_container, ticTacToe)
-                .commit();
-        */
-        // If we saved any Instructions, we need to recreate the board state.
-        if(savedInstanceState != null) {
-            mTurn = savedInstanceState.getBoolean(mKEY_TURN);
-            mInstructions = savedInstanceState.getStringArrayList(mKEY_INSTRUCTIONS);
-            for(int i = 0; i < mInstructions.size(); i++) {
-                sendMessage(mInstructions.get(i));
-            }
-            // If there is an odd number of Instructions, then the turn indicator will be set to
-            // the wrong value. We can handle this here.
-            if(mInstructions.size() % 2 != 0) {
-                mTurn = !mTurn;
-            }
-            //ticTacToe.checkNotFinished();
-        }
-
     }
 
     @Override
@@ -211,37 +165,13 @@ public class MainActivity extends AppCompatActivity
         return true;
     }
 
-    @Override public void onSaveInstanceState(final Bundle outState) {
-        super.onSaveInstanceState(outState);
-
-        // Store the Instructions and Turn for the save state.
-        outState.putStringArrayList(mKEY_INSTRUCTIONS, mInstructions);
-        outState.putBoolean(mKEY_TURN, mTurn);
-    }
-
     /**
      * Sends a message alerting the event handling system that the new game button was clicked.
      *
      * @param view the new game button.
      */
     public void onNewGame(final View view) {
-        // Create the message.
-        String msg = getTurn() + "\n";
-        msg += view.getTag().toString();
-
-        // Empty the instructions list, as a new game has begun.
-        mInstructions.clear();
-        mInstructions.add(msg);
-
-        // Output New Game Message
-        String newTurn = (getTurn().equals(getString(R.string.player_1)) ?
-                "Player 1 (" + getString(R.string.xValue) + ")" :
-                "Player 2 (" + getString(R.string.oValue) + ")") + "'s Turn";
-        Snackbar start = Snackbar.make(findViewById(R.id.activity_main), "New Game! " + newTurn, Snackbar.LENGTH_SHORT);
-        start.show();
-
-        //TODO: replace this with an implemented event handling system.
-        sendMessage(msg);
+        ((TTTFragment) Panel.ttt.getFragment(getApplicationContext())).onNewGame(view);
     }
 
     /** The FAB click handler.  Show the various options. */
@@ -257,40 +187,10 @@ public class MainActivity extends AppCompatActivity
      * @param view the tile clicked
      */
     public void tileOnClick(final View view) {
-        String msg = getTurn() + "\n";
-        msg = msg + view.getTag().toString();
-
-        // Keep track of mInstructions for recreating the board.
-        mInstructions.add(msg);
-
-        //TODO: replace this with an implemented event handling system.
-        sendMessage(msg);
-
-        mTurn = !mTurn;
+        ((TTTFragment) Panel.ttt.getFragment(getApplicationContext())).tileOnClick(view);
     }
 
     // Private instance methods.
-
-    /**
-     * A placeholder method for a message handler / event coordinator to be implemented at a later time.
-     *
-     * @param msg the message to transmit to the message handler.
-     */
-    private void sendMessage(final String msg) {
-        //TODO: replace this with an implemented event handling system.
-        // This will be a switch for each of the individual game fragment handlers.
-        ticTacToe.messageHandler(msg);
-    }
-
-    /**
-     * Gets the current mTurn and returns a string reflecting the player's
-     * name who is currently playing.
-     *
-     * @return player 1 or player 2, depending on the mTurn.
-     */
-    private String getTurn() {
-        return mTurn ? mPLAYER1 : mPLAYER2;
-    }
 
     // Private nested classes
 
@@ -307,12 +207,12 @@ public class MainActivity extends AppCompatActivity
         /**
          * Build an adapter to handle the panels.
          *
-         * @param fm The fragment manager.
+         * @param manager The fragment manager.
          */
         public GameChatPagerAdapter(final FragmentManager manager) {
             super(manager);
             panelList.add(Panel.chat);
-            panelList.add(Panel.game);
+            panelList.add(Panel.ttt);
         }
 
         @Override
